move en-US strings in react-components
